Move newGlobals to Hydra.Node.Run
This is the only usage site. Another viable location would be
hydra-cardano-api.
